Move event subscriptions to top level viewmodel
Much of the major narration events for the entire page were handled by the AudioWorkspace, which, should simply be responsible for the visualization and marker interactions. Many such events effect the page as a whole, and thus, need to have data propagated downward to the other ViewModels.
